+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/responses-pilot/format/page.tsx
@@ -2,7 +2,7 @@ import { Metadata } from "next";
 import { SelectFormat } from "./SelectFormat";
 import { ApiClientGuard } from "../guards/ApiClientGuard";
 import { LocationGuard } from "../guards/LocationGuard";
-import { getPageTitle } from "../lib/getStepOf";
+import { getPageTitle } from "../lib/getPageTitle";
 
 export async function generateMetadata(props: {
   params: Promise<{ locale: string }>;
