+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/responses-pilot/lib/getPageTitle.ts
@@ -1,0 +1,27 @@
+import { serverTranslation } from "@i18n";
+import { steps, getStepOf } from "./getStepOf";
+
+export async function getPageTitle({
+  step,
+  props,
+}: {
+  step?: keyof typeof steps;
+  props: Promise<{ locale: string }>;
+}): Promise<{ title: string }> {
+  const params = await props;
+  const { locale } = params;
+  const { t } = await serverTranslation("response-api", { lang: locale });
+
+  if (!step || !steps[step]) {
+    return {
+      title: `${t("section-title")}`,
+    };
+  }
+
+  const { current, total } = getStepOf(step);
+  const title = `${t("section-title")} — ${t("stepOf", { current, total })}`;
+
+  return {
+    title,
+  };
+}
