+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/responses-pilot/load-key/page.tsx
@@ -1,6 +1,6 @@
 import { Metadata } from "next";
 import { SelectApiKey } from "./SelectApiKey";
-import { getPageTitle } from "../lib/getStepOf";
+import { getPageTitle } from "../lib/getPageTitle";
 
 export async function generateMetadata(props: {
   params: Promise<{ locale: string }>;
